Move whitelist test out of `tests.scala`
